// src/app/(DashboardLayout)/types/apps/Expert.ts
export interface ExpertType {
  id: number;
  name: string;
  phone: string;
  email: string;
  job: string;
  about: string;
  specialization: string;
  practical_experiences: string;
  training_courses: string;
  academic_qualifications: string;
  research: string;
  nationality: string;
  resident_country: string;


}